refactor: decompose exercises, exercise-form, muscle-volume, exercise-group-card (batch 4)
Extract hooks and sub-components from 4 large files:
- exercises.tsx (503→239 lines): ExerciseCard, ExerciseDetailPane
- ExerciseForm.tsx (476→312 lines): useExerciseForm, MuscleGroupPicker
- MuscleVolumeSegment.tsx (471→241 lines): useMuscleVolume, VolumeBarChart, VolumeTrendChart
- ExerciseGroupCard.tsx (463→237 lines): GroupCardHeader, SuggestionChip
